splitting preprocessing and training scripts
