+++ jobs/slurm_inference.sh
@@ -10,8 +10,8 @@
 #SBATCH --output=/anvil/scratch/x-mho1/jobout/%x_%A_%a.out  # Output file for each array task
 #SBATCH --error=/anvil/scratch/x-mho1/jobout/%x_%A_%a.out   # Error file for each array task
 
-# SLURM_ARRAY_TASK_ID=0
-export TQDM_DISABLE=0
+SLURM_ARRAY_TASK_ID=5
+# export TQDM_DISABLE=0
 
 module restore cmass
 conda activate cmass
@@ -25,14 +25,14 @@ nbody=mtnglike
 sim=fastpm
 infer=default
 
-halo=True
+halo=False
 galaxy=False
-ngc=False
+ngc=True
 sgc=False
 mtng=False
 
 extras="nbody.zf=0.500015"
-device=cuda
+device=cpu
 
 postfix="nbody=$nbody sim=$sim infer=$infer infer.exp_index=$exp_index"
 postfix="$postfix infer.halo=$halo infer.galaxy=$galaxy"
@@ -40,4 +40,5 @@ postfix="$postfix infer.ngc_lightcone=$ngc infer.sgc_lightcone=$sgc infer.mtng_l
 postfix="$postfix infer.device=$device $extras"
 
 echo "Running inference with $postfix"
+python -m cmass.infer.preprocess $postfix
 python -m cmass.infer.train $postfix
